Fix most warnings (#35)
* Turn on warnings

* -Wno-unused-parameter -Wno-unused-variable -Wno-unused-function

* Fix -Wformat

```
warning: format specifies type 'unsigned long long' but the argument has type 'uint64_t' (aka 'unsigned long') [-Wformat]
```

* Fix -Wrange-loop-construct warning

```
warning: loop variable 'it' creates a copy from type 'std::pair<const std::basic_string<char>, RT64::PresetLights::Light> const' [-Wrange-loop-construct]
```

* -Wsign-compare

```
comparison of integers of different signs: 'int32_t' (aka 'int') and 'uint32_t' (aka 'unsigned int')
```

* -Wrange-loop-construct

```
warning: loop variable ‘it’ creates a copy from type ‘const std::pair<const std::__cxx11::basic_string<char>, RT64::PresetDrawCall>’
```

* rerun

* Fix -Wparentheses

```
warning: suggest parentheses around arithmetic in operand of ‘^’
```

* -Wsign-compare

```
warning: comparison of integer expressions of different signedness: ‘int32_t’ {aka ‘int’} and ‘unsigned int’
```

* Update hlslpp to fix a number of warnings

* Remove `#pragma once` in cpp file

```
warning: #pragma once in main file
```

* -Wswitch

```
warning: enumeration value ‘None’ not handled in switch
```

* -Wformat-security

```
warning: format not a string literal and no format arguments
```

* regen

* Move warnings to clang check

* Fix -Wdelete-non-abstract-non-virtual-dtor

```
warning: destructor called on non-final 'RT64::PresetMaterial' that has virtual functions but non-virtual destructor
```

* remove warnings file

* Fix one unused parameter

* Fix assert wrapping

* Add -Werror flag

* Change type of `dstIndex` to `uint32_t`

* More uses I missed

* Fix new warnings

+++ src/gui/rt64_debugger_inspector.cpp
@@ -261,7 +261,7 @@ namespace RT64 {
         uint32_t totalSyncCount = 0;
         for (uint32_t f = 0; f < workload.fbPairCount; f++) {
             auto &fbPair = workload.fbPairs[f];
-            if (renderer.framebufferIndex == f) {
+            if ((uint32_t)renderer.framebufferIndex == f) {
                 renderer.framebufferAddress = renderer.framebufferDepth ? fbPair.depthImage.address : fbPair.colorImage.address;
             }
 
@@ -896,7 +896,7 @@ namespace RT64 {
                                     ImGui::PushID(l);
                                     
                                     if (openCall) {
-                                        bool open = (openLoadIndex == (callDesc.loadIndex + l));
+                                        bool open = ((uint32_t)openLoadIndex == (callDesc.loadIndex + l));
                                         ImGui::SetNextItemOpen(open);
                                         if (open) {
                                             ImGui::SetScrollHereY(0.0f);
@@ -984,7 +984,7 @@ namespace RT64 {
                                     const DrawCallTile &callTile = drawData.callTiles[callDesc.tileIndex + t];
                                     const LoadTile &loadTile = callTile.loadTile;
                                     if (openCall) {
-                                        bool open = (openTileIndex == (callDesc.tileIndex + t));
+                                        bool open = ((uint32_t)openTileIndex == (callDesc.tileIndex + t));
                                         ImGui::SetNextItemOpen(open);
                                         if (open) {
                                             ImGui::SetScrollHereY(0.0f);
