+++ src/hle/rt64_rsp.h
@@ -232,9 +232,9 @@ namespace RT64 {
         void computeModelViewProj();
         void specialComputeModelViewProj();
         void setModelViewProjChanged(bool changed);
-        void setVertex(uint32_t address, uint8_t vtxCount, uint8_t dstIndex);
-        void setVertexPD(uint32_t address, uint8_t vtxCount, uint8_t dstIndex);
-        void setVertexEXV1(uint32_t address, uint8_t vtxCount, uint8_t dstIndex);
+        void setVertex(uint32_t address, uint8_t vtxCount, uint32_t dstIndex);
+        void setVertexPD(uint32_t address, uint8_t vtxCount, uint32_t dstIndex);
+        void setVertexEXV1(uint32_t address, uint8_t vtxCount, uint32_t dstIndex);
         void setVertexColorPD(uint32_t address);
         template<bool addEmptyVelocity>
         void setVertexCommon(uint8_t dstIndex, uint8_t dstMax);
